release 1.0

+++ dvd-screensaver/Item.cpp
@@ -3,15 +3,15 @@
 #include <cmath>
 
 void Item::Update(
-    float delta,
+    double delta,
     float camL, float camR,
     float camB, float camT
 )
 {
     HitX = HitY = false;
 
-    float dX = XDir * Speed * delta;
-    float dY = YDir * Speed * delta;
+    double dX = XDir * Speed * delta;
+    double dY = YDir * Speed * delta;
 
     X += dX;
     if (X - Width * 0.5f < camL || X + Width * 0.5f > camR)
